+++ build.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 
-rm hashtopolis.zip
+if [ -f hashtopolis.zip ]; then
+  rm hashtopolis.zip
+fi
 zip -r hashtopolis.zip __main__.py htpclient -x "*__pycache__*"