+++ htpclient/chunk.py
@@ -35,6 +35,8 @@ class Chunk:
                 return -2
             elif ans['status'] == 'fully_dispatched':
                 return 0
+            elif ans['status'] == 'health_check':
+                return -3
             else:
                 self.chunk = ans
                 return 1
@@ -48,10 +50,11 @@ class Chunk:
         if ans is None:
             logging.error("Failed to send keyspace!")
             sleep(5)
-            return 0
+            return False
         elif ans['response'] != 'SUCCESS':
             logging.error("Sending of keyspace failed: " + str(ans))
             sleep(5)
-            return 0
+            return False
         else:
             logging.info("Keyspace got accepted!")
+            return True