+++ htpclient/jsonRequest.py
@@ -1,18 +1,19 @@
 import logging
-import requests
 
 from htpclient.config import *
+from htpclient.session import *
 
 
 class JsonRequest:
     def __init__(self, data):
         self.data = data
         self.config = Config()
+        self.session = Session().s
 
     def execute(self):
         try:
             logging.debug(self.data)
-            r = requests.post(self.config.get_value('url'), json=self.data)
+            r = self.session.post(self.config.get_value('url'), json=self.data)
             if r.status_code != 200:
                 logging.error("Status code from server: " + str(r.status_code))
                 return None